% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_plot.R
\name{decision_plot}
\alias{decision_plot}
\title{decision_plot}
\usage{
decision_plot(
  object,
  obs_num,
  title = NULL,
  geo.unit = NULL,
  geo.id = NULL,
  bar_default = TRUE
)
}
\arguments{
\item{object}{Enter the name of the object that contains the model's contributions
and results obtained using the explain function.}

\item{obs_num}{single or multiple observation numbers}

\item{title}{plot title}

\item{geo.unit}{The name of the stratum variable in the BARP model as a character.}

\item{geo.id}{Enter a single value of the stratum variable as a character.}
}
\value{
plot_out The waterfall plot of the observation at index obs_num.
}
\description{
The decision_plot function is a graph that visualizes how individual features
contribute to a model's prediction for a specific observation using Shapley values.
It can be used to visualize one or multiple observations.
}
