% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.barp.R
\name{explain.barp}
\alias{explain.barp}
\title{explain.barp}
\usage{
\method{explain}{barp}(
  object,
  feature_names = NULL,
  X = NULL,
  nsim = 1,
  pred_wrapper = NULL,
  census = NULL,
  geo.unit = NULL,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{object}{A BARP model (Bayesian Additive Regression Tree) estimated
using the barps function, a modified version of the barp function from the BARP library with a fixed seed.}

\item{feature_names}{The name of the variable for which you want to check the contribution.
The default value is set to NULL, which means the contribution of all variables in X will be calculated.}

\item{X}{The dataset containing all independent variables used as input when estimating the BART model.}

\item{nsim}{The number of MCMC simulation iterations, which is fixed at 1 by default
in the case of the BARP model.}

\item{pred_wrapper}{A function used to estimate the predicted values of the model.}

\item{parallel}{The default value is set to FALSE,
but it can be changed to TRUE for parallel computation.}

\item{newdata}{New data containing the variables included in the model.
This is used when checking the contribution of newly input data using the model.
The default value is set to NULL, meaning that the input X data,
i.e., the data used for model estimation, will be used by default.}
}
\value{
phis A list containing the Shapley values for each variable.

newdata The data used to check the contribution of variables.
If a variable has two categories, it is dummy-coded, and if it has three or more categories,
categorical variables are one-hot encoded.

fnull The expected value of the model's predictions.

fx The prediction value for each observation.

factor_names The name of the categorical variable.
If the data contains only continuous or dummy variables, it is set to NULL.
}
\description{
This function is implemented to calculate the contribution of each variable
in the BARP(Bayesian Additive Regression Tree with post-stratification) model
using the permutation method.
}
\references{
Štrumbelj, Erik, and Igor Kononenko. "Explaining prediction models and individual
predictions with feature contributions." Knowledge and information systems 41 (2014): 647-665.
Bisbee, James. "Barp: Improving mister p using bayesian additive regression trees." American Political Science Review 113.4 (2019): 1060-1065.
}
