% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_column_default.R
\name{explain_column_default}
\alias{explain_column_default}
\title{explain_column_default}
\usage{
explain_column_default(object, X, column, pred_wrapper, newdata)
}
\description{
A modified function for handling categorical variables.
}
