% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_default.R
\name{plot.explain}
\alias{plot.explain}
\title{plot}
\usage{
\method{plot}{explain}(
  object,
  average = NULL,
  type = NULL,
  num_post = NULL,
  plot.flag = TRUE,
  adjust = FALSE,
  probs = 0.95,
  title = NULL,
  ...
)
}
\arguments{
\item{object}{explainBART object}

\item{...}{add}
}
\value{
Return plot:
\item{out}{A \code{plot} plot of Mean (|SHAP|)}
}
\description{
This function uses XXXXXXX
}
