% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall_plot}
\alias{waterfall_plot}
\title{waterfall_plot}
\usage{
waterfall_plot(object, obs_num, title = NULL, geo.unit = NULL, geo.id = NULL)
}
\arguments{
\item{object}{Enter the name of the object that contains the model's contributions
and results obtained using the explain function.}

\item{obs_num}{observation number (only one)}

\item{title}{plot title}

\item{geo.unit}{The name of the stratum variable in the BARP model as a character.}

\item{geo.id}{Enter a single value of the stratum variable as a character.}
}
\value{
plot_out The waterfall plot of the observation at index obs_num.
}
\description{
The waterfall_plot function is a bar chart that displays the positive and
negative contributions across sequential data points, visualizing how each
variable's contributions change for a single observation.
}
