% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_explainBART.R
\name{plot.ExplainBART}
\alias{plot.ExplainBART}
\alias{plot}
\title{A Function for Visualizing the Shapley Values of BART Models}
\usage{
\method{plot}{ExplainBART}(
  x,
  average = NULL,
  type = NULL,
  num_post = NULL,
  plot.flag = TRUE,
  adjust = FALSE,
  probs = 0.95,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{ExplainBART} class object containing the Shapley values of the BART model.}

\item{average}{Input the reference value for calculating the mean of the object's \code{phi} list.
\code{"obs"} represents abind the average based on observations (#post by #variable),
while \code{"post"} represents the average based on posterior samples (#obs by #variable).
If \code{"both"} is entered, calculations are performed based on both observation and posterior sample criteria.
If no value is specified, "both" is used as the default.}

\item{type}{\code{"bar"} represents a bar chart that includes the average contribution of each variable,
while \code{"bee"} represents a summary plot, allowing you to determine the graph's format.}

\item{num_post}{To check the contribution of variables for a single posterior sample,
enter a value within the number of posterior samples.}

\item{plot.flag}{If \code{average = "obs"}, the quantile interval of each variable's is provided by default.}

\item{adjust}{The default value is \code{FALSE}.
Enter \code{TRUE} to check the Shapley values adjusted based on the model's average contribution.}

\item{probs}{Enter the probability for the quantile interval. The default value is \code{0.95}.}

\item{title}{The title of the plot, with a default value of \code{NULL}.}

\item{xlab}{Enter the label to be displayed on the x-axis. If not provided, a default label will be used.}

\item{ylab}{Enter the label for the y-axis if needed.}

\item{...}{Additional arguments to be passed}
}
\value{
The plot is returned based on the specified option.: 
\item{out}{If average is \code{"obs"} or \code{"post"}, a bar plot or summary plot is generated based on the selected averaging criterion. 
When \code{average} is set to \code{"both"}, either a bar plot or a boxplot comparing the distributions of Shapley values computed under the two averaging criteria is generated.
In the case where a boxplot is produced, the observation-based and posterior-sample-based summaries can additionally be rendered separately
via \code{out$observation} and \code{out$post}, respectively.
If adjust is \code{TRUE}, the adjusted Shapley values are displayed.
If \code{num_post} is specified, a bar plot or summary plot for the selected posterior sample is generated.}
}
\description{
The \code{plot.ExplainBART} function provides various visualization methods for Shapley values. 
It is designed to visualize \code{ExplainBART} class objects, which contain Shapley values computed from models estimated using the \code{bart} function from the \pkg{dbarts} or the \code{wbart}/\code{gbart} functions from \pkg{BART}.
The values and format used in the graph are determined based on the input parameters.
}
\examples{
\donttest{
## Friedman data
set.seed(2025)
n <- 200
p <- 5
X <- data.frame(matrix(runif(n * p), ncol = p))
y <- 10 * sin(pi* X[ ,1] * X[,2]) +20 * (X[,3] -.5)^2 + 10 * X[ ,4] + 5 * X[,5] + rnorm(n)

## Using dbarts 
model <- dbarts::bart (X,y, keeptrees = TRUE, ndpost = 200)

# prediction wrapper function
pfun <- function (object, newdata) {
predict(object, newdata)
}

# Calculate shapley values
model_exp <-  Explain  ( model, X = X,  pred_wrapper =  pfun )

# Distribution of Shapley values (boxplot) 
# computed based on observation and posterior sample criteria
plot(model_exp,average = "both" ) 

# Barplot based on observation criteria
plot(model_exp,average = "obs",type ="bar",probs = 0.95)

# Barplot based on posterior sample
plot(model_exp,average = "post",type ="bar" )

# Summary plot based on posterior sample
plot(model_exp,average = "post",type ="bees" )

# Summary plot of the 100th posterior sample
plot(model_exp,average = "post",type ="bees",num_post = 100) 

# Barplot of the adjusted baseline
plot(model_exp, type ="bar",  adjust= TRUE )
}
}
